browser/devtools/webide/test/test_runtime.html

<!DOCTYPE html>

<html>

  <head>
    <meta charset="utf8">
    <title></title>

    <script type="application/javascript" src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
    <script type="application/javascript" src="chrome://mochikit/content/chrome-harness.js"></script>
    <script type="application/javascript;version=1.8" src="head.js"></script>
    <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css">
  </head>

  <body>

    <script type="application/javascript;version=1.8">
      window.onload = function() {
        SimpleTest.waitForExplicitFinish();

        Task.spawn(function* () {

          function isPlayActive() {
            return !win.document.querySelector("#cmd_play").hasAttribute("disabled");
          }

          function isStopActive() {
            return !win.document.querySelector("#cmd_stop").hasAttribute("disabled");
          }

          Cu.import("resource://gre/modules/devtools/dbg-server.jsm");
          DebuggerServer.init(function () { return true; });
          DebuggerServer.addBrowserActors();

          let win = yield openWebIDE();

          win.AppManager.runtimeList.usb.push({
            connect: function(connection) {
              ok(connection, win.AppManager.connection, "connection is valid");
              connection.host = null; // force connectPipe
              connection.connect();
              return promise.resolve();
            },

            getName: function() {
              return "fakeRuntime";
            }
          });

          win.AppManager.update("runtimelist");

          let packagedAppLocation = getTestFilePath("app");

          yield win.Cmds.importPackagedApp(packagedAppLocation);

          let panelNode = win.document.querySelector("#runtime-panel");
          let items = panelNode.querySelectorAll(".runtime-panel-item-usb");
          is(items.length, 1, "Found one runtime button");

          let deferred = promise.defer();
          win.AppManager.connection.once(
              win.Connection.Events.CONNECTED,
              () => deferred.resolve());

          items[0].click();

          ok(win.document.querySelector("window").className, "busy", "UI is busy");
          yield win.UI._busyPromise;

          is(Object.keys(DebuggerServer._connections).length, 1, "Connected");

          ok(isPlayActive(), "play button is enabled 1");
          ok(!isStopActive(), "stop button is disabled 1");
          let oldProject = win.AppManager.selectedProject;
          win.AppManager.selectedProject = null;

          yield nextTick();

          ok(!isPlayActive(), "play button is disabled 2");
          ok(!isStopActive(), "stop button is disabled 2");
          win.AppManager._selectedProject = oldProject;
          win.UI.updateCommands();

          yield nextTick();

          ok(isPlayActive(), "play button is enabled 3");
          ok(!isStopActive(), "stop button is disabled 3");


          yield win.Cmds.disconnectRuntime();

          is(Object.keys(DebuggerServer._connections).length, 0, "Disconnected");

          ok(win.AppManager.selectedProject, "A project is still selected");
          ok(!isPlayActive(), "play button is disabled 4");
          ok(!isStopActive(), "stop button is disabled 4");

          deferred = promise.defer();
          win.AppManager.connection.once(
              win.Connection.Events.CONNECTED,
              () => deferred.resolve());

          win.document.querySelectorAll(".runtime-panel-item-custom")[1].click();

          yield deferred.promise;

          is(Object.keys(DebuggerServer._connections).length, 1, "Locally connected");

          yield win.Cmds.disconnectRuntime();

          yield closeWebIDE(win);

          DebuggerServer.destroy();

          yield removeAllProjects();

          SimpleTest.finish();

        });
      }


    </script>
  </body>
</html>
